Added Mouse Enemy
Created mouse sprite and object and it's variable definitions. Then added basic movement, wandering back and forth. Added two states for the mouse; a wandering state and a stand off state. If the player is within a certain distance of the mouse, the mouse will stop moving and face the direction of the player.

THINGS TO ADD
- attack
- fix moonwalk glitch