StrayCat State 1 and scripts
Added state 2 to the stray cat enemy. It now chases after the player if it is within it's range as well as in its territory.

Also added a couple of scripts to clean up enemy coding. I added a movement script for the stray cat, and a face player script for general enemy use.